requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
attrs==25.4.0
blinker==1.9.0
bond==4.1.2
build==1.3.0
buildozer==1.5.0
builtin-interfaces==1.2.2
cartographer-ros-msgs==2.0.9002
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
composition-interfaces==1.2.2
control-msgs==4.8.0
controller-manager==2.51.0
controller-manager-msgs==2.51.0
cv-bridge==3.2.1
data-tamer-msgs==0.9.3
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
diagnostic-updater==4.0.6
distlib==0.4.0
domain-coordinator==0.10.0
dwb-msgs==1.1.18
dynamixel-sdk==3.8.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.12.19
gazebo-msgs==3.9.0
generate-parameter-library-py==0.5.0
geometry-msgs==4.9.0
gunicorn==23.0.0
humanfriendly==10.0
idna==3.11
image-geometry==3.2.1
ImageIO==2.37.2
interactive-markers==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark-parser==0.12.0
laser-geometry==2.4.0
launch==1.0.10
launch-param-builder==0.1.1
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lazy_loader==0.4
lifecycle-msgs==1.2.2
llvmlite==0.46.0
logging-demo==0.20.5
map-msgs==2.1.0
MarkupSafe==3.0.3
message-filters==4.3.8
moveit-configs-utils==2.5.9
moveit-msgs==2.2.1
mpmath==1.3.0
nav-2d-msgs==1.1.18
nav-msgs==4.9.0
nav2-common==1.1.18
nav2-msgs==1.1.18
nav2-simple-commander==1.0.0
networkx==3.4.2
numba==0.63.1
numpy==2.2.6
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
osrf-pycommon==2.1.6
packaging==25.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==12.0.0
platformdirs==4.5.0
plotjuggler-msgs==0.2.3
pooch==1.8.2
protobuf==6.33.2
PyMatting==1.1.14
pyproject_hooks==1.2.0
python-qt-binding==1.1.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2controlcli==2.51.0
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpds-py==0.30.0
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scikit-image==0.25.2
scipy==1.15.3
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
sh==1.14.3
shape-msgs==4.9.0
slam-toolbox==2.6.10
smclib==4.1.2
srdfdom==2.0.7
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.14.0
teleop-twist-keyboard==2.4.0
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
tifffile==2025.5.10
tomli==2.3.0
topic-monitor==0.20.5
tqdm==4.67.1
trajectory-msgs==4.9.0
turtlebot3-applications-msgs==1.0.1
turtlebot3-autorace-camera==1.2.2
turtlebot3-autorace-detect==1.2.2
turtlebot3-autorace-mission==1.2.2
turtlebot3-example==2.3.3
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.3.3
turtlesim==1.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
ur-msgs==2.3.0
urdfdom-py==1.2.1
urllib3==2.6.2
virtualenv==20.35.4
visualization-msgs==4.9.0
Werkzeug==3.1.4
xacro==2.0.13
